+++ src/zodbshootout/_dbsupport.py
@@ -132,15 +132,21 @@ class MappingFactory(object):
 
         db = DB(MappingStorage())
         import transaction
+
         db.close = lambda: None
         self.data.populate(lambda: db)
         del db.close
+
         mconn = db.open()
         mroot = mconn.root()
         for worker in range(self.concurrency):
             mroot['speedtest'][worker].update(self.data.data_to_store())
         transaction.commit()
         mconn.cacheMinimize()
+        # Clear transfer counts so as not to mess up any downstream
+        # assertions about what they should be. (No other database does
+        # stuff on opening.)
+        mconn.getTransferCounts(True)
         mconn.close()
         return db