+++ src/zodbshootout/_runner.py
@@ -234,6 +234,7 @@ def _run_benchmarks_for_contender(runner, options, data, db_factory):
             ('%s: read %d cold objects', speedtest.bench_cold_read, 'cold',),
             ('%s: read %d cold prefeteched objects', speedtest.bench_cold_read_prefetch,
              'prefetch_cold',),
+            ('%s: readCurrent %d objects', speedtest.bench_readCurrent, 'readCurrent',),
             ('%s: read %d warm objects', speedtest.bench_read_after_write, 'warm',),
             ('%s: read %d hot objects', speedtest.bench_hot_read, 'hot',),
             ('%s: read %d steamin objects', speedtest.bench_steamin_read, 'steamin',),